Add better wait-for-idle mechanism to STL perf tests
This CL adds a better way to wait for the STL animation to finish from
within the UIautomator performance tests.

Before this CL, we would wait for the fromScene to disappear, but it was
not really correct as UIAutomator will consider the fromScene to
disappear as soon as it was fully occluded by the toScene. Moreover, for
b/395625468 we want to add new performance tests for overlays and motion
mechanics, so waiting for the fromScene to disappear is not possible
anymore with overlay transitions.

Bug: 395625468
Test: Ran the benchmarks locally
Flag: EXEMPT test only
Change-Id: I3274eeeffc44d304554ddc4b15e730fd2c46a984