Keep AudioInjector audio ports open
The audio ports from the remote submix are connected when the source
audio track is playing. They are needed to be available when the
permission checks are done for the remote device.
Created always playing source audio tracks that play silence when no
audio data from the remote audio device is available.

Bug: 381386115
Test: manual recording using the Vdm sample apps
Flag: EXEMPT, demo sample app only
Change-Id: I69db2da27b545c3cada87420cd0202c75da7a29d